refactor(tests): streamline tooltip tests with reusable hover utility
- Replaced inline hover logic in Tooltip and Card tests with a new `testHoverTooltip` utility function for better code reuse and readability.
- Updated imports to utilize the new utility, enhancing test maintainability and consistency across components.
- Added the `testHoverTooltip` function to handle hover interactions and tooltip assertions in a centralized manner.
